Add files via upload
sword strikes added, dash is there again. I apologize for the lack of polish regarding the color and pants, this is more for mechanical testing. I'll clean it up later it's just a long and monotonous process. At the very very end I will smooth everything over, just want to do that when I'm sure the exact poses won't need to be redone.